+++ packages/core/split-pane/CHANGELOG.md
@@ -3,6 +3,66 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.3.3](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/split-pane@0.3.2...@kong-ui-public/split-pane@0.3.3) (2026-01-24)
+
+**Note:** Version bump only for package @kong-ui-public/split-pane
+
+
+
+
+
+## [0.3.2](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/split-pane@0.3.1...@kong-ui-public/split-pane@0.3.2) (2026-01-24)
+
+
+### Bug Fixes
+
+* **deps:** update all non-major dependencies with stable versions ([#2854](https://github.com/Kong/public-ui-components/issues/2854)) ([8db6a9b](https://github.com/Kong/public-ui-components/commit/8db6a9bfffb91a72a5c436cf2cfb8e2c5fdb66a2))
+
+
+
+
+
+## [0.3.1](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/split-pane@0.3.0...@kong-ui-public/split-pane@0.3.1) (2026-01-24)
+
+
+### Bug Fixes
+
+* **deps:** update dependency @kong/icons to ^1.42.0 ([#2857](https://github.com/Kong/public-ui-components/issues/2857)) ([bb65b9e](https://github.com/Kong/public-ui-components/commit/bb65b9eea7489142268aee86d5cdb6de338762f9))
+
+
+
+
+
+# [0.3.0](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/split-pane@0.2.0...@kong-ui-public/split-pane@0.3.0) (2026-01-23)
+
+
+### Features
+
+* **split-pane:** add SplitToolbar component [KHCP-19305] ([#2851](https://github.com/Kong/public-ui-components/issues/2851)) ([1a0dc42](https://github.com/Kong/public-ui-components/commit/1a0dc42092f105f763902a7b0aaaf970f2e9deab))
+
+
+
+
+
+# [0.2.0](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/split-pane@0.1.3...@kong-ui-public/split-pane@0.2.0) (2026-01-22)
+
+
+### Features
+
+* **split-pane:** add splitpane component [KHCP-19160] ([#2823](https://github.com/Kong/public-ui-components/issues/2823)) ([50ea9ca](https://github.com/Kong/public-ui-components/commit/50ea9ca485b0f3dd42cba678f7784652fbb6e0a4))
+
+
+
+
+
+## [0.1.3](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/split-pane@0.1.2...@kong-ui-public/split-pane@0.1.3) (2026-01-21)
+
+**Note:** Version bump only for package @kong-ui-public/split-pane
+
+
+
+
+
 ## [0.1.2](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/split-pane@0.1.1...@kong-ui-public/split-pane@0.1.2) (2026-01-19)
 
 **Note:** Version bump only for package @kong-ui-public/split-pane