+++ packages/entities/entities-keys/CHANGELOG.md
@@ -3,6 +3,63 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.11.3](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-keys@3.11.2...@kong-ui-public/entities-keys@3.11.3) (2026-01-24)
+
+**Note:** Version bump only for package @kong-ui-public/entities-keys
+
+
+
+
+
+## [3.11.2](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-keys@3.11.1...@kong-ui-public/entities-keys@3.11.2) (2026-01-24)
+
+
+### Bug Fixes
+
+* **deps:** update all non-major dependencies with stable versions ([#2854](https://github.com/Kong/public-ui-components/issues/2854)) ([8db6a9b](https://github.com/Kong/public-ui-components/commit/8db6a9bfffb91a72a5c436cf2cfb8e2c5fdb66a2))
+
+
+
+
+
+## [3.11.1](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-keys@3.11.0...@kong-ui-public/entities-keys@3.11.1) (2026-01-24)
+
+
+### Bug Fixes
+
+* **deps:** update dependency @kong/icons to ^1.42.0 ([#2857](https://github.com/Kong/public-ui-components/issues/2857)) ([bb65b9e](https://github.com/Kong/public-ui-components/commit/bb65b9eea7489142268aee86d5cdb6de338762f9))
+
+
+
+
+
+# [3.11.0](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-keys@3.10.46...@kong-ui-public/entities-keys@3.11.0) (2026-01-22)
+
+
+### Features
+
+* **entities-routes, entities-keys:** prepopulate name when create entity ([#2846](https://github.com/Kong/public-ui-components/issues/2846)) ([cbfefc1](https://github.com/Kong/public-ui-components/commit/cbfefc13488fc804a42c0dff65bb9289043ba7db))
+
+
+
+
+
+## [3.10.46](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-keys@3.10.45...@kong-ui-public/entities-keys@3.10.46) (2026-01-21)
+
+**Note:** Version bump only for package @kong-ui-public/entities-keys
+
+
+
+
+
+## [3.10.45](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-keys@3.10.44...@kong-ui-public/entities-keys@3.10.45) (2026-01-21)
+
+**Note:** Version bump only for package @kong-ui-public/entities-keys
+
+
+
+
+
 ## [3.10.44](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-keys@3.10.43...@kong-ui-public/entities-keys@3.10.44) (2026-01-20)
 
 **Note:** Version bump only for package @kong-ui-public/entities-keys